Break out the main function of glimpse into its own file
